+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+#3.1.3
+- Small fixes and improvements
+
 #3.1.2
 - Implemented 'Convert time' button
 - Selecton is capable now of converting astrix punctuation for feet/inches (like 6'5")